+++ documentation/playground/src/molecules/AutocompleteExamples.tsx
@@ -34,7 +34,7 @@ export function AutocompleteExamples() {
     <div className="u-margin" style={{ width: '700px' }}>
       <h2>Autocomplete</h2>
       <div className="u-margin">
-        <Autocomplete id="autocomplete-example-1" label="Select National Parc (Uncontrolled)" items={items} />
+        <Autocomplete required id="autocomplete-example-1" label="Select National Parc (Uncontrolled)" items={items} />
         <Autocomplete
           id="autocomplete-example-2"
           noResultsText="NOTHING FOUND"