+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - An `Alert` component without a title label will no longer add a margin pushing the description down.
+- Can use the `required` property for the `Autocomplete` component.
+- Fixed `asterisk` typos in documentation.
 
 =======